requirements.txt
dash==2.8.1
dash_core_components==2.0.0
dash_html_components==2.0.0
pandas==1.4.4
Pillow==9.5.0
plotly==5.9.0
gunicorn
openpyxl
gspread
